python-backend/face_recognition/requirements.txt
# Echoplex Face Recognition Backend Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Face Recognition
# Note: dlib installation on Windows requires Visual C++ Build Tools
# Try: pip install dlib-binary  (pre-built wheel, easier installation)
# Or install Visual C++ Build Tools from Microsoft
face-recognition==1.3.0
# dlib will be installed via face-recognition dependencies
# If dlib fails, try: pip install dlib-binary

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.26.4

# Optional: Advanced detection (if needed)
# ultralytics==8.1.0  # YOLOv8 for person detection
# torch==2.2.0

# Utilities
pydantic==2.5.0

